Refactor countdown into GameInfo
